refactor(footer): Extract footer into it's own component
Isolates it so it can be changed/tested easier.
