+++ __tests__/App.test.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { shallow } from "enzyme";
-import App from "./App";
+
+import App from "../src/App";
 
 it("renders without crashing", () => {
   shallow(<App />);
